add /sql & some notes

+++ src/main/java/com/airconditioner/community/dto/CommentCreateDTO.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigInteger;
+
 /**
  * @Author AirConditioner
  * @Date 2022/9/11 20:26
@@ -12,7 +14,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class CommentCreateDTO {
-    private int parentId;
+    private BigInteger parentId;
     private String content;
     private Integer type;
 }
