+++ src/main/java/com/airconditioner/community/service/CommentService.java
@@ -1,10 +1,11 @@
 package com.airconditioner.community.service;
 
-import com.airconditioner.community.bean.Comment;
-import com.airconditioner.community.bean.User;
+import com.airconditioner.community.entity.Comment;
+import com.airconditioner.community.entity.User;
 import com.airconditioner.community.dto.CommentDTO;
 import com.airconditioner.community.enums.CommentTypeEnum;
 
+import java.math.BigInteger;
 import java.util.List;
 
 /**
@@ -12,7 +13,13 @@ import java.util.List;
  * @Date 2022/9/11 20:52
  **/
 public interface CommentService {
+    /**
+     * 创建评论
+     *
+     * @param comment (parentId, content, type, gmtModified, gmtCreate, commentator)
+     * @param user    ()
+     */
     void insert(Comment comment, User user);
 
-    List<CommentDTO> listByTargetId(Integer id, CommentTypeEnum type);
+    List<CommentDTO> listByTargetId(BigInteger id, CommentTypeEnum type);
 }
